💦 Refactor: AuthUtil 클래스 생성을 통한 리팩토링
