+++ src/main/java/com/otclub/humate/domain/auth/util/AuthUtil.java
@@ -1,4 +1,6 @@
-package com.otclub.humate.domain.auth.constant;
+package com.otclub.humate.domain.auth.util;
+
+import jakarta.servlet.http.Cookie;
 
 /**
  * Auth 모듈에서 사용할 상수 클래스
@@ -12,7 +14,7 @@ package com.otclub.humate.domain.auth.constant;
  * 2024.07.30  	조영욱        최초 생성
  * </pre>
  */
-public class AuthConstant {
+public class AuthUtil {
     // JWT Token
     public static final long ACCESS_TOKEN_EXPIRE_TIME = 600_000L; // 10분 (millisecond)
     public static final long REFRESH_TOKEN_EXPIRE_TIME = 604_800_000L; // 1주 (millisecond)
@@ -20,4 +22,12 @@ public class AuthConstant {
     // JWT Cookie
     public static final int JWT_TOKEN_COOKIE_MAX_AGE = 604_800; // 1주 (second)
 
+    public static Cookie createJwtTokenCookie(String key, String value) {
+        Cookie cookie = new Cookie(key, value);
+        cookie.setMaxAge(JWT_TOKEN_COOKIE_MAX_AGE);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+
+        return cookie;
+    }
 }
